feat(util): refactor out useCloseOnDocumentEvents hook
